+++ src/hap.ts
@@ -21,7 +21,11 @@ import { LockMechanism } from './types/lock-mechanism';
 import { SecuritySystem } from './types/security-system';
 import { Switch } from './types/switch';
 import { Television } from './types/television';
+import { ContactSensor } from './types/contact-sensor';
+import { OccupancySensor } from './types/occupancy-sensor';
+import { MotionSensor } from './types/motion-sensor';
 import { TemperatureSensor } from './types/temperature-sensor';
+import { Battery } from './types/battery-status';
 import { Thermostat } from './types/thermostat';
 import { Window } from './types/window';
 import { WindowCovering } from './types/window-covering';
@@ -42,6 +46,8 @@ export class Hap {
 
   public ready: boolean;
 
+  private dummy = () => {};
+  
   /* GSH Supported types */
   types = {
     Door: new Door(),
@@ -55,11 +61,16 @@ export class Hap {
     Outlet: new Switch('action.devices.types.OUTLET'),
     SecuritySystem: new SecuritySystem(),
     Switch: new Switch('action.devices.types.SWITCH'),
-    Television: new Television(),
+    Television: new Television(this),
     TemperatureSensor: new TemperatureSensor(this),
     Thermostat: new Thermostat(this),
     Window: new Window(),
     WindowCovering: new WindowCovering(),
+    Speaker: this.dummy,
+    InputSource: this.dummy,
+    ContactSensor: new ContactSensor(),
+    OccupancySensor: new OccupancySensor(),
+    Battery: new Battery(),
   };
 
   /* event tracking */
@@ -89,6 +100,14 @@ export class Hap {
     Characteristic.CurrentRelativeHumidity,
     Characteristic.SecuritySystemTargetState,
     Characteristic.SecuritySystemCurrentState,
+    Characteristic.ActiveIdentifier,
+    Characteristic.Mute,
+    Characteristic.ContactSensorState,
+    Characteristic.OccupancyDetected,
+    Characteristic.CurrentMediaState,
+    Characteristic.MotionDetected,
+    Characteristic.StatusLowBattery,
+    Characteristic.BatteryLevel,
   ];
 
   instanceBlacklist: Array<string> = [];
@@ -194,12 +213,14 @@ export class Hap {
    * Build Google SYNC intent payload
    */
   async buildSyncResponse(): Promise<SmartHomeV1SyncDevices[]> {
-    const devices = this.services.map((service) => {
-      if (!this.types[service.type]) {
-        // this.log.debug(`Unsupported service type ${service.type}`);
-        return;
-      }
-      // console.log('buildSyncResponse', service);
+    const devices = this.services.filter((service) =>
+      this.types?.[service.type]?.sync,
+    ).map((service) => {
+      // if (!this.types[service.type]) {
+      //   // this.log.debug(`Unsupported service type ${service.type}`);
+      //   return;
+      // }
+      // // console.log('buildSyncResponse', service);
       return this.types[service.type].sync(service);
     });
     return devices;
@@ -251,7 +272,7 @@ export class Hap {
     for (const command of commands) {
       for (const device of command.devices) {
         const service = this.services.find(x => x.uniqueId === device.id);
-        this.log.debug(`Processing command ${command.execution[0].command} for ${device.id} and ${service}`);
+        this.log.debug(`Processing command ${command.execution[0].command} for ${device.id} and ${service.serviceName}`);
         if (service) {
           // check if two factor auth is required, and if we have it
           if (this.config.twoFactorAuthPin && this.types[service.type].twoFactorRequired
@@ -325,10 +346,24 @@ export class Hap {
       }
       services = services.filter(x => this.types[x.type] !== undefined);
       this.log.debug(`Loaded ${services.length} accessories from Homebridge - pre filter`);
+      services = services.filter(x => !this.instanceBlacklist.find(y => y === x?.instance?.username));
+      // Pre-compile accessoryFilter strings into RegExp objects
+      const compiledAccessoryFilter = this.accessoryFilter.map(filter => new RegExp(filter));
+      const searchList = (target: string, regexList: RegExp[]): boolean => {
+        if (target) {
+          for (const regex of regexList) {
+            if (regex.test(target)) {
+              this.log.debug(`${this.accessoryFilterInverse ? 'Including' : 'Skipping'} service '${target}' - matches accessoryFilter '${regex}'`);
+              return true;
+            }
+          }
+        }
+        return false;
+      };
       if (this.accessoryFilterInverse) {
-        services = services.filter(x => this.accessoryFilter.includes(x.serviceName));
+        services = services.filter(x => searchList(x.serviceName, compiledAccessoryFilter));
       } else {
-        services = services.filter(x => !this.accessoryFilter.includes(x.serviceName));
+        services = services.filter(x => !searchList(x.serviceName, compiledAccessoryFilter));
       }
       services = services.filter(x => !this.accessorySerialFilter.includes(x.accessoryInformation['Serial Number']));
       // if 2fa is forced for this service type, but a pin has not been set ignore the service
@@ -388,6 +423,9 @@ export class Hap {
 
     for (const uniqueId of pendingStateReport) {
       const service = this.services.find(x => x.uniqueId === uniqueId);
+      if (!this.types?.[service.type]?.query) {
+        continue;
+      }
       states[service.uniqueId] = this.types[service.type].query(service);
     }
 
@@ -401,10 +439,12 @@ export class Hap {
     if (!this.services.length) {
       return;
     }
-    this.services.map((service) => {
-      if (!this.types[service.type]) {
-        return;
-      }
+    this.services.filter((service) => 
+      this.types?.[service.type]?.query,
+    ).map((service) => {
+      // if (!this.types[service.type]) {
+      //   return;
+      // }
       return states[service.uniqueId] = this.types[service.type].query(service);
     });
     return await this.sendStateReport(states);